fix inconsistent indentation in README
It was a mix of tabs and spaces, and the tabs were obviously meant to
be interpreted as being 4 columns wide, which is a much less common
interpretation than 8 columns.  So make the file self-consistent and
unambiguous by replacing each tab with 4 spaces.

Then fix a couple of issues with indentation in the example
docker-compose file.