Set restart policy to unless-stopped
If a host is rebooted gracefully or via crash, typically users will
want the docker containers to automatically restart.  Otherwise Mautic
will become unavailable until the next time a human notices the issue
and manually restarts the containers, and that could risk an
unacceptably long outage period during which Mautic is completely
broken.