Fix docker-compose syntax error
the docker-compose had multiple syntax and indentation errors, it was missing the version and services block and part of the mautic service parameters where indented with tabs, also the database link had an unnecessary name tag and the database host env var was referring to localhost instead of the DB service linked to it. Also there is a volume linked to a data folder that doesn't apply to most cases and is not referred to in this readme, if a "plug and play" installation is intended, is best to use the data folder in the image and explain, if data persistence is needed, then other configs must be performed.