feat: Read db password from file and check startup parameters
The entrypoint for the apache web image was improved and now checks that the mounted volumes are present, writeable by the www-data user and that the environment variables MAUTIC_DB_HOST, MAUTIC_DB_PORT, MAUTIC_DB_USER and MAUTIC_DB_PASSWORD are present.

If one of the conditions is not met, it will give an error message to the console, to give new users a feedback on their docker configuration.

For working with docker secrets, the new environment variable MAUTIC_DB_PASSWORD_FILE was added. It can point to a file in the container to read the MAUTIC_DB_PASSWORD from.

All changes to the entrypoint are secured with tests. The tests are located in the tests directory and can be run with the run_entry_point_web_tests.sh script. The script uses the shunit2 test library. See  https://github.com/kward/shunit2/raw/refs/heads/master/shunit2 for details.